requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==9.5.0
sqlite3